app/requirements.txt
flask
tldextract
python-whois
google-generativeai # Added this line for the Gemini API
flask
bcrypt
blinker
click
itsdangerous
Jinja2
MarkupSafe
Werkzeug
gunicorn
python-dotenv
Flask
Flask-PyMongo
Flask-Session
flask_session 